+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Add LineThickness property to TrackerControl (#1831)
 - Add properties for `MinimumSegmentLength` to series and annotations (#1853)
 - Add fractal examples for PolygonAnnotation and PolylineAnnotations (#1853)
+- Add `AxisPreference` to `PlotManipulator`
 
 ### Changed
 - Update net40 and net45 to net452 (#1835)